Fix compiler warnings in shell

+++ apps/shell/shell.c
@@ -532,7 +532,7 @@ shell_strtolong(const char *str, const char **retstr)
     ++strptr;
   }
   
-  for(i = 0; i < 10 && isdigit(strptr[i]); ++i) {
+  for(i = 0; i < 10 && isdigit((int)strptr[i]); ++i) {
     num = num * 10 + strptr[i] - '0';
   }
   if(retstr != NULL) {
